+++ ExternalConsole/Config/PluginConfig.cs
@@ -5,6 +5,7 @@ using BepInEx.Configuration;
 public static class PluginConfig {
   public static ConfigEntry<bool> IsModEnabled { get; private set; }
   public static ConfigEntry<string> ExternalInputFilename { get; private set; }
+  public static ConfigEntry<string> CommandLogFilename { get; private set; }
 
   public static void BindConfig(ConfigFile config) {
     IsModEnabled =
@@ -20,5 +21,12 @@ public static class PluginConfig {
             "externalInputFilename",
             "external-input.txt",
             "Filename for the text file to use for external input.");
+
+    CommandLogFilename =
+        config.Bind(
+            "ExternalInput",
+            "commandLogFilename",
+            "ec-command-log.txt",
+            "Filename for text-file to log all processed commands from external-input.");
   }
 }